requirements.txt
SpeechRecognition
pyaudio
websockets
aiofiles
google-genai
anthropic
openai
python-dotenv
